Update es.xml with the new translations
I know there is a pr suggesting this changes, #567, but it has way too many spelling mistakes, so I made a new pr. I stuck the most I could with Minecraft's original translations.   
Also, in line 242, it can be both "la" and "el" (before "{3}"), referring to the sword and the bow, respectively. I left "la", as the most common object for knocking a player is a sword, but something would have to be done, so feel free to suggest.   
I also corrected a tittle in which the Spanish academy changed recently, to make it even.
I feel atm this pr can be merged.